PLT-1429 Extend prototype to all V2 builtins functions in UPLC  (#5267)
The formalisation was extended to add support for pairs, lists, and data.
All V2 builtin functions are now supported.
All new functionality was extended to the plc-agda executable, for UPLC (execution mode U).

In order to test the new functionality some tests in the conformance tests (for pairs and lists) with expected result "parser error", were fixed so that they work with the current parser. No test concerning data was modified, since the current parser only supports data in serialised form.