[Errors] Swap operational and structural errors (#6439)
This simply swaps around two type variables and two constructors, affecting `EvaluationError` and `EvaluationException`. Structural errors correspond to type errors and operational errors correspond to runtime errors, hence the former "precede" the latter. This PR finally reflects that perceived ordering. It's not really important, but I've got things mixed up before due to the previously counter-intuitive ordering, so I thought it'd be worth fixing it, given how easy that is.