[Builtins] [Test] Dump denotation signatures to golden files (#6462)
This dumps the type signature of the denotation of each of the builtins to a golden file. All changes in type signatures must be reflected explicitly whenever they're captured by semantics variants (see the test code). This is a very important test given how easy it is to change the behavior of a builtin just by slightly tweaking the signature of its denotation.